Update UI components for routing
