+++ ticketing.ui/src/Login.tsx
@@ -1,34 +1,15 @@
 import React from 'react'; 
-import { makeStyles } from '@material-ui/core/styles';
-import TextField from '@material-ui/core/TextField';
-import Button from '@material-ui/core/Button';
 
 import Header from './Header' ;
-import history from './History';
+import LoginForm from './LoginForm';
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    '& > *': {
-      margin: theme.spacing(1),
-      width: '25ch',
-    },
-    display: "grid",
-    justifyItems: "center",
-    margin: "10%",
-  },
-}));
-
-export default function Login() : JSX.Element {
-    const classes = useStyles();
-  
+export default class App extends React.Component {
+  render() {
     return (
-      <>
-        <Header /> 
-          <form className={classes.root} noValidate autoComplete="off">
-              <TextField id="text-username" label="Username" />
-              <TextField id="text-password" label="Password" type="password" />
-              <Button onClick={() => history.push('/Dashboard')}>Login</Button>
-          </form>
+      <> 
+        <Header />
+        <LoginForm />
       </>
     );
   }
+}
